Split up the Text.Heterocephalus.Parse module.
